Renamed KonaClassic to kona_classic
