+++ keyboards/kona_classic/keymaps/ansi/config.h
