+++ keyboards/kona_classic/keymaps/ansi/rules.mk
