+++ keyboards/kona_classic/keymaps/ansi_split_arrows/keymap.c
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "KonaClassic.h"
+#include "kona_classic.h"
 
 #define MODS_SHFT_MASK  (MOD_BIT(KC_LSHIFT)|MOD_BIT(KC_RSHIFT)|MOD_BIT(KC_LGUI)|MOD_BIT(KC_RGUI))
 #define MODS_GUI_MASK   (MOD_BIT(KC_LGUI)|MOD_BIT(KC_RGUI))
@@ -32,8 +32,8 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 		KC_F1, KC_F2,  F(0),	KC_1,    KC_2,     KC_3,   KC_4,    KC_5,    KC_6,    KC_7,   KC_8,    KC_9,    KC_0,    KC_MINS, KC_EQL,  KC_BSPC, \
 		KC_F3, KC_F4,  KC_TAB,	KC_Q,    KC_W,     KC_E,   KC_R,    KC_T,    KC_Y,    KC_U,   KC_I,    KC_O,    KC_P,    KC_LBRC, KC_RBRC, KC_BSLS, \
 		KC_F5, KC_F6,  KC_CAPS, KC_A,    KC_S,     KC_D,   KC_F,    KC_G,    KC_H,    KC_J,   KC_K,    KC_L,    KC_SCLN, KC_QUOT, KC_ENT,  ______, \
-		KC_F7, KC_F8,  KC_LSFT, ______,  KC_Z,     KC_X,   KC_C,    KC_V,    KC_B,    KC_N,   KC_M,    KC_COMM, KC_DOT,  KC_SLSH, KC_RSFT,   ______, \
-		KC_F9, KC_F10, KC_LCTL, KC_LGUI, KC_LALT,  KC_ENT,          MO(_FN), KC_SPC,                   KC_LALT, KC_RGUI, ______,  MO(_FN), KC_RCTL
+		KC_F7, KC_F8,  KC_LSFT, ______,  KC_Z,     KC_X,   KC_C,    KC_V,    KC_B,    KC_N,   KC_M,    KC_COMM, KC_DOT,  KC_SLSH, KC_UP,   MO(_FN), \
+		KC_F9, KC_F10, KC_LCTL, KC_LGUI, KC_LALT,  KC_ENT,          MO(_FN), KC_SPC,                   KC_RALT, KC_RCTL, KC_LEFT, KC_DOWN, KC_RGHT
 	),
 	[_FN] = KEYMAP(
 		RGB_TOG, RGB_MOD, KC_TILD, ______, ______, ______, ______, ______, ______, ______, ______, ______, ______, ______, ______, RESET, \
